Vector2: move pybind template helper to .cc file
This isolates the template code into its own
translation unit.
